Change Connection.Write() and Connection.Read()
This commit changes the behaviour of Read() and Write of Connection.
Before, the input array was split at every "\n" and then sent in chunks.
Afterwards, the "\n" was added on the receiving side back again - even
if the input didn't end with a "\n", a "\n" was added.

The changes in this commit "escape" all instances of "\n" with "\\n". On
the reading side, those changes are reverted. The caller now must only
make sure that no "\\n" occurs in the input, because it would be
transformed into "\n".
This commit fixes #1.